app/src/test/java/com/example/simpleparadox/listycity/CityListTest.java
package com.example.simpleparadox.listycity;

import org.junit.jupiter.api.Test;


import static org.junit.jupiter.api.Assertions.*;

class CityListTest {

    private CityList mockCityList() {
        CityList cityList = new CityList();
        cityList.add(mockCity());
        return cityList;
    }

    private City mockCity() {
        return new City("Edmonton", "Alberta");
    }

    @Test
    void testAdd() {
        CityList cityList = mockCityList();
        assertEquals(1, cityList.getCities().size());

        City city = new City("Regina", "Saskatchewan");
        cityList.add(city);
        assertEquals(2, cityList.getCities().size());

        assertTrue(cityList.getCities().contains(city));        // asserts that condition=true i.e city is added.
    }

    @Test
    void testAddException() {
        CityList cityList = mockCityList();

        City city = new City("Yellowknife", "Northwest Territories");
        cityList.add(city);

        assertThrows(IllegalArgumentException.class, () -> {
            cityList.add(city);
        });
    }

    @Test
    void testGetCities()
    {
        CityList cityList = mockCityList();
        assertEquals(0, mockCity().compareTo(cityList.getCities().get(0)));

        City city = new City("Charlottetown", "Prince Edward Island");
        cityList.add(city);
        assertEquals(0, city.compareTo(cityList.getCities().get(0)));

        assertEquals(0, mockCity().compareTo(cityList.getCities().get(1)));
    }

    @Test
    void testCountCities()
    {
        CityList cityList = mockCityList();
        assertEquals(1, cityList.getCities().size());

        City city = new City("Regina", "Saskatchewan");
        cityList.add(city);

        assertNotNull(cityList);        // asserts that cityList is not null

        assertEquals(2, cityList.CountCities());    // asserts that total size is now = 2
    }

    @Test
    void testDeleteCity()
    {
        CityList cityList = mockCityList();
        assertEquals(1, cityList.getCities().size());

        City city = new City("Regina", "Saskatchewan");
        cityList.add(city);
        assertEquals(2, cityList.getCities().size());

        cityList.DeleteCity(city);
        assertEquals(1, cityList.getCities().size());   // after deletion, size = size -1

        assertFalse(cityList.getCities().contains(city));     // asserts that the city is deleted from list.
    }

}